_W_ funs and buffer struct, find_esc -> read_next
Major re-write to ensure all writing functions get the full
struct buffer for.

Also, removed the parallel logic in find_esc because the
risk of forgetting about it and letting it get out of
sync with read_esc is just too high (we almost forgot for
the 1.0 release and OSC changes).